requirements.txt
requests
flask
flask_cors
google-genai
cohere
IPython
soundfile
gradio
scipy
python-dotenv